pyproject.toml
[project]
name = "benchmarkrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.7",
    "faiss-cpu>=1.13.1",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-groq>=1.1.1",
    "numpy>=2.3.5",
    "psutil>=7.1.3",
    "pymupdf>=1.26.7",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
